+++ webapp/src/components/ConfidenceHistogramTopWords.tsx
@@ -10,16 +10,25 @@ import {
 } from "services/api";
 import TopWords from "components/TopWords/TopWords";
 import TopWordsSkeleton from "components/TopWords/TopWordsSkeleton";
-import { QueryFilterState, QueryPipelineState, WordCount } from "types/models";
+import {
+  QueryFilterState,
+  QueryPaginationState,
+  QueryPipelineState,
+  WordCount,
+} from "types/models";
 import { ALL_OUTCOMES } from "utils/const";
 
 type Props = {
+  baseUrl: string;
   filters: QueryFilterState;
+  pagination: QueryPaginationState;
   pipeline: Required<QueryPipelineState>;
 };
 
 const ConfidenceHistogramTopWords: React.FC<Props> = ({
+  baseUrl,
   filters,
+  pagination,
   pipeline,
 }) => {
   const { jobId, datasetSplitName } = useParams<{
@@ -88,7 +97,14 @@ const ConfidenceHistogramTopWords: React.FC<Props> = ({
         {isFetchingTopWords ? (
           <TopWordsSkeleton />
         ) : (
-          <TopWords wordCounts={correctWordCounts} palette="success" />
+          <TopWords
+            baseUrl={baseUrl}
+            filters={filters}
+            pagination={pagination}
+            pipeline={pipeline}
+            wordCounts={correctWordCounts}
+            palette="success"
+          />
         )}
         <Typography display="inline" align="center" variant="caption">
           Counts of most {topWords?.importanceCriteria ?? "important"} words for
@@ -97,7 +113,14 @@ const ConfidenceHistogramTopWords: React.FC<Props> = ({
         {isFetchingTopWords ? (
           <TopWordsSkeleton />
         ) : (
-          <TopWords wordCounts={errorWordCounts} palette="error" />
+          <TopWords
+            baseUrl={baseUrl}
+            filters={filters}
+            pagination={pagination}
+            pipeline={pipeline}
+            wordCounts={errorWordCounts}
+            palette="error"
+          />
         )}
       </Box>
     </Box>